+++ AdColony/src/main/java/com/mopub/mobileads/AdColonyInterstitial.java
@@ -3,22 +3,16 @@ package com.mopub.mobileads;
 import android.app.Activity;
 import android.content.Context;
 import android.os.Handler;
-import android.text.TextUtils;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.adcolony.sdk.AdColony;
-import com.adcolony.sdk.AdColonyAppOptions;
 import com.adcolony.sdk.AdColonyInterstitialListener;
 import com.adcolony.sdk.AdColonyZone;
-import com.mopub.common.MoPub;
 import com.mopub.common.logging.MoPubLog;
-import com.mopub.common.privacy.ConsentStatus;
-import com.mopub.common.privacy.PersonalInfoManager;
 import com.mopub.common.util.Json;
 
-import java.util.Arrays;
 import java.util.Map;
 
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
@@ -31,34 +25,14 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
 
 public class AdColonyInterstitial extends CustomEventInterstitial {
-    /*
-     * We recommend passing the AdColony client options, app ID, all zone IDs, and current zone ID
-     * in the serverExtras Map by specifying Custom Event Data in MoPub's web interface.
-     *
-     * Please see AdColony's documentation for more information:
-     * https://github.com/AdColony/AdColony-Android-SDK-3
-     */
-    private static final String DEFAULT_CLIENT_OPTIONS = "version=YOUR_APP_VERSION_HERE,store:google";
-    private static final String DEFAULT_APP_ID = "YOUR_AD_COLONY_APP_ID_HERE";
-    private static final String[] DEFAULT_ALL_ZONE_IDS = {"ZONE_ID_1", "ZONE_ID_2", "..."};
-    private static final String DEFAULT_ZONE_ID = "YOUR_CURRENT_ZONE_ID";
-    private static final String CONSENT_RESPONSE = "consent_response";
-    private static final String CONSENT_GIVEN = "explicit_consent_given";
 
-    /*
-     * These keys are intended for MoPub internal use. Do not modify.
-     */
-    public static final String CLIENT_OPTIONS_KEY = "clientOptions";
-    public static final String APP_ID_KEY = "appId";
-    public static final String ALL_ZONE_IDS_KEY = "allZoneIds";
-    public static final String ZONE_ID_KEY = "zoneId";
-    public static final String ADAPTER_NAME = AdColonyInterstitial.class.getSimpleName();
+    private static final String ADAPTER_NAME = AdColonyInterstitial.class.getSimpleName();
 
     private CustomEventInterstitialListener mCustomEventInterstitialListener;
     private AdColonyInterstitialListener mAdColonyInterstitialListener;
     private final Handler mHandler;
     private com.adcolony.sdk.AdColonyInterstitial mAdColonyInterstitial;
-    private static String[] previousAdColonyAllZoneIds;
+
     @NonNull
     private AdColonyAdapterConfiguration mAdColonyAdapterConfiguration;
 
@@ -73,89 +47,52 @@ public class AdColonyInterstitial extends CustomEventInterstitial {
                                     @Nullable Map<String, Object> localExtras,
                                     @NonNull Map<String, String> serverExtras) {
         if (!(context instanceof Activity)) {
-            customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
             MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(), MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
             return;
         }
 
-        String clientOptions = DEFAULT_CLIENT_OPTIONS;
-        String appId = DEFAULT_APP_ID;
-        String[] allZoneIds = DEFAULT_ALL_ZONE_IDS;
-        String zoneId = DEFAULT_ZONE_ID;
+        String clientOptions = serverExtras.get(AdColonyAdapterConfiguration.CLIENT_OPTIONS_KEY);
+        if (clientOptions == null)
+            clientOptions = "";
 
         mCustomEventInterstitialListener = customEventInterstitialListener;
 
-        if (extrasAreValid(serverExtras)) {
-            clientOptions = serverExtras.get(CLIENT_OPTIONS_KEY);
-            appId = serverExtras.get(APP_ID_KEY);
-            allZoneIds = extractAllZoneIds(serverExtras);
-            zoneId = serverExtras.get(ZONE_ID_KEY);
-            mAdColonyAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
+
+        // Set mandatory parameters
+        final String appId = AdColonyAdapterConfiguration.getAdColonyParameter(AdColonyAdapterConfiguration.APP_ID_KEY, serverExtras);
+        final String zoneId = AdColonyAdapterConfiguration.getAdColonyParameter(AdColonyAdapterConfiguration.ZONE_ID_KEY, serverExtras);
+
+        String[] allZoneIds;
+        String allZoneIdsString = AdColonyAdapterConfiguration.getAdColonyParameter(AdColonyAdapterConfiguration.ALL_ZONE_IDS_KEY, serverExtras);
+        if (allZoneIdsString != null) {
+            allZoneIds = Json.jsonArrayToStringArray(allZoneIdsString);
+        } else {
+            allZoneIds = null;
         }
 
-        // Check to see if app ID parameter is present. If not AdColony will not return an ad.
-        // So there's no need to make a request. If so, must fail and log the flow.
-        if (TextUtils.isEmpty(appId) || TextUtils.equals(appId, DEFAULT_APP_ID)) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "AppId parameter cannot be empty. " +
-                    "Please make sure you enter correct AppId on the MoPub Dashboard " +
-                    "for AdColony.");
-
-            customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+        // Check if mandatory parameters are valid, abort otherwise
+        if (appId == null) {
+            abortRequestForIncorrectParameter(AdColonyAdapterConfiguration.APP_ID_KEY);
             return;
         }
 
-        AdColonyAppOptions mAdColonyAppOptions = null;
-        if (!TextUtils.isEmpty(clientOptions)) {
-            mAdColonyAppOptions = AdColonyAppOptions.getMoPubAppOptions(clientOptions);
+        if (zoneId == null || allZoneIds == null || allZoneIds.length == 0) {
+            abortRequestForIncorrectParameter(AdColonyAdapterConfiguration.ZONE_ID_KEY);
+            return;
         }
 
-        // Pass the user consent from the MoPub SDK to AdColony as per GDPR
-        PersonalInfoManager personalInfoManager = MoPub.getPersonalInformationManager();
-
-        boolean canCollectPersonalInfo = MoPub.canCollectPersonalInformation();
-        boolean shouldAllowLegitimateInterest = MoPub.shouldAllowLegitimateInterest();
-
-        mAdColonyAppOptions = mAdColonyAppOptions == null ? new AdColonyAppOptions() :
-                mAdColonyAppOptions;
-
-        if (personalInfoManager != null && personalInfoManager.gdprApplies() == Boolean.TRUE) {
-            if (shouldAllowLegitimateInterest) {
-                if (personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.EXPLICIT_NO
-                        || personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.DNT) {
-                    mAdColonyAppOptions.setOption(CONSENT_GIVEN, true)
-                            .setOption(CONSENT_RESPONSE, false);
-                } else {
-                    mAdColonyAppOptions.setOption(CONSENT_GIVEN, true)
-                            .setOption(CONSENT_RESPONSE, true);
-                }
-            } else {
-                mAdColonyAppOptions.setOption(CONSENT_GIVEN, true)
-                        .setOption(CONSENT_RESPONSE, canCollectPersonalInfo);
-            }
-        }
-
+        mAdColonyAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
         mAdColonyInterstitialListener = getAdColonyInterstitialListener();
-        if (!isAdColonyConfigured()) {
-            if (!TextUtils.isEmpty(appId)) {
-                AdColony.configure((Activity) context, mAdColonyAppOptions, appId, allZoneIds);
-            }
-        } else if ((shouldReconfigure(previousAdColonyAllZoneIds, allZoneIds))) {
-            // Need to check the zone IDs sent from the MoPub portal and reconfigure if they are
-            // different than the zones we initially called AdColony.configure() with
-            if (!TextUtils.isEmpty(appId)) {
-                AdColony.configure((Activity) context, mAdColonyAppOptions, appId, allZoneIds);
-            }
-            previousAdColonyAllZoneIds = allZoneIds;
-        } else {
-            // If state of consent has changed and we aren't calling configure again, we need
-            // to pass this via setAppOptions()
-            AdColony.setAppOptions(mAdColonyAppOptions);
-        }
 
-        if (!TextUtils.isEmpty(zoneId)) {
-            AdColony.requestInterstitial(zoneId, mAdColonyInterstitialListener);
-            MoPubLog.log(zoneId, LOAD_ATTEMPTED, ADAPTER_NAME);
-        }
+        AdColonyAdapterConfiguration.checkAndConfigureAdColonyIfNecessary(context, clientOptions, appId, allZoneIds);
+        AdColony.requestInterstitial(zoneId, mAdColonyInterstitialListener);
+        MoPubLog.log(zoneId, LOAD_ATTEMPTED, ADAPTER_NAME);
+    }
+
+    private void abortRequestForIncorrectParameter(String parameterName) {
+        AdColonyAdapterConfiguration.logAndFail("interstitial request", parameterName);
+        mCustomEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
     }
 
     @Override
@@ -178,14 +115,10 @@ public class AdColonyInterstitial extends CustomEventInterstitial {
     @Override
     protected void onInvalidate() {
         if (mAdColonyInterstitial != null) {
-            mAdColonyInterstitialListener = null;
             mAdColonyInterstitial.destroy();
             mAdColonyInterstitial = null;
         }
-    }
-
-    private boolean isAdColonyConfigured() {
-        return !AdColony.getSDKVersion().isEmpty();
+        mAdColonyInterstitialListener = null;
     }
 
     private AdColonyInterstitialListener getAdColonyInterstitialListener() {
@@ -252,41 +185,5 @@ public class AdColonyInterstitial extends CustomEventInterstitial {
                 }
             };
         }
-    }
-
-    private boolean extrasAreValid(Map<String, String> extras) {
-        return extras != null
-                && extras.containsKey(CLIENT_OPTIONS_KEY)
-                && extras.containsKey(APP_ID_KEY)
-                && extras.containsKey(ALL_ZONE_IDS_KEY)
-                && extras.containsKey(ZONE_ID_KEY);
-    }
-
-    private static boolean shouldReconfigure(String[] previousZones, String[] newZones) {
-        // If AdColony is configured already, but previousZones is null, then that means AdColony
-        // was configured with the AdColonyRewardedVideo adapter so attempt to configure with
-        // the ids in newZones. They will be ignored within the AdColony SDK if the zones are
-        // the same as the zones that the other adapter called AdColony.configure() with.
-        if (previousZones == null) {
-            return true;
-        } else if (newZones == null) {
-            return false;
-        } else if (previousZones.length != newZones.length) {
-            return true;
-        }
-        Arrays.sort(previousZones);
-        Arrays.sort(newZones);
-        return !Arrays.equals(previousZones, newZones);
-    }
-
-    private String[] extractAllZoneIds(Map<String, String> serverExtras) {
-        String[] result = Json.jsonArrayToStringArray(serverExtras.get(ALL_ZONE_IDS_KEY));
-
-        // AdColony requires at least one valid String in the allZoneIds array.
-        if (result.length == 0) {
-            result = new String[]{""};
-        }
-
-        return result;
     }
 }